Add pg_basebackup flag to force overwrite of destination data directory
Currently, pg_basebackup has a hard restriction where the destination
data directory must be empty or nonexistant. It is expected that
anything of interest should be moved somewhere temporarily and then
copied back in. To reduce the complexity, we introduce a new flag
--force-overwrite which will delete the directories or files that are
being copied from the source data directory before doing the actual
copy. Combined with the Greenplum-specific exclusion flag (-E), we are
now able to preserve files of interest.

Our main example would be gprecoverseg full recovery and pg_log
files. There have been times when a mirror fails and a full recovery
would run which would drop the entire mirror directory before running
pg_basebackup which would result in the mirror log files before the
crash to be erased. This is substantially worse when we think of
gprecoverseg rebalancing scenario where we currently do not have
pg_rewind and must run full recovery to bring the old primary back
up... which would result in vast amounts of old primary log files to
be erased. Then during rebalance, the acting primary which would
return to being a mirror also goes through a full recovery so its logs
as a primary are also removed. The obvious solution would be to tar
these logs out and untar them back in afterwards, but what if there
are other files that must be preserved. Creating a copy may be costly
in environments where disk space is valued highly.