Invalidate AppendOnlyHash cache	entry at end of TRUNCATE
TRUNCATE will rewrite the relation by creating a temporary table and
swapping it with the real relation. For AO, this includes the
auxiliary tables which is concerning for the AO relation's pg_aoseg
table which holds information that a AO segment file is available for
write or waiting to be compacted/dropped. Since we do not currently
invalidate the AppendOnlyHash cache entry, the entry could have
invisible leaks in its AOSegfileStatus array that will be stuck in
state AOSEG_STATE_AWAITING_DROP. These leaks will persist until the
user evicts the cache entry by not using the table to allow another AO
table to cache itself in that slot or by restarting the database. We
fix this issue by invalidating the cache entry at the end of TRUNCATE
on AO relations.

Conflicts:
    src/backend/commands/tablecmds.c
     - TRUNCATE is a bit different between Greenplum 5 and 6. Needed
       to move the heap_close() to the end to invalidate the
       AppendOnlyHash entries after dispatch.
     - Macros RelationIsAppendOptimized() and IS_QUERY_DISPATCHER() do
       not exist in 5X_STABLE.
    src/test/isolation2/sql/truncate_after_ao_vacuum_skip_drop.sql
     - Isolation2 utility mode connections use dbid instead of content
       id like in Greenplum 6.