Modify a sublink test VIEW to avoid unnecessary diffs in pg_upgrade test
When the v_xpect_triangle_de VIEW from sublink regress test is dumped and
restored, an unnecessary alias is added to one variable in a UNION which causes
a syntactic difference but not a semantic one. This results in an unnecessary
diff during pg_upgrade testing even though the functionality of the VIEW is the
same. We do not want to change the underlying logic so we modify the VIEW so
that no diff is unnecessarily produced for pg_upgrade tests.

Example:
From the test SQL file...
(SELECT 1 AS "offset", 'HT' AS triangle UNION SELECT 0 AS "offset", 'LT') t

From the \d+ of the VIEW which is what is dumped on old cluster...
(SELECT 1 AS "offset", 'HT'::text AS triangle UNION SELECT 0 AS "offset", 'LT'::text) t

From the \d+ of the VIEW after restoring the VIEW on new cluster...
(SELECT 1 AS "offset", 'HT'::text AS triangle UNION SELECT 0 AS "offset", 'LT'::text AS text) t

With this patch, it will show in dump and after restore as...
(SELECT 1 AS "offset", 'HT'::text AS triangle UNION SELECT 0 AS "offset", 'LT'::text AS triangle) t

Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>