Autogenerate recovery.conf file at the end of pg_rewind
Like pg_basebackup, we should generate recovery.conf file at the end of
pg_rewind so that utilities do not have to take care of that step. This is
required in Greenplum mainly because users will not use pg_rewind manually. Note
that we only autogenerate recovery.conf file if pg_rewind is called with source
server because we utilize the libpq connection information. We expect pg_rewind
usage to only be through gprecoverseg.

Added TODO message to create common library between pg_basebackup and pg_rewind
to create the recovery.conf file since most of this code addition is copied from
pg_basebackup.c file (there are a couple diffs to make it work for pg_rewind).

Co-authored-by: Paul Guo <pguo@pivotal.io>