Fix simple_no_rewind_required pg_rewind intermittent test failure
The old master can be slow to restart its WAL stream with promoted standby as
part of automatic timeline switch for the new timeline after finishing old
timeline catchup. The SELECT query on pg_stat_replication was executed too fast
and the entry would show state STARTUP which is the default state for timeline
switching. To make the test more deterministic, loop the check on
pg_stat_replication for STREAMING state with timeout 30 seconds.