Add MASTER ONLY option for LOCK TABLE
Currently, LOCK TABLE will always dispatch the lock statement to the
segments. However, the dispatch is unnecessary in many common cases
(e.g. when you just want to prevent DDL changes from ALTER TABLE or
DROP TABLE). Having the ability to not dispatch would give a small
performance boost in large Greenplum clusters when explicitly locking
a large amount of tables in a transaction.

To skip the dispatch, we introduce a new option to LOCK TABLE called
MASTER ONLY.  When added to the LOCK TABLE query, the lock will only
be taken on the master.