Enable Hot Standby
This enables hot standby for the standby master and allows the standby
master to dispatch read-only queries to the HA mirrors (as established
in gp_segment_configuration). Currently, read queries work and don't
seem to be hindered by ongoing transactions on the master and primary
segments... probably because we use synchronous WAL replication for
our HA.

To enable hot standby for the cluster, the master and primary segments
must have wal_level=hot_standby in the postgresql.conf. Afterwards,
the mirrors and standby master can be created via pg_basebackup (or if
you have existing mirrors, you can crash the mirrors with SIGTERM and
then running gprecoverseg -F to recreate all the mirrors... standby
master needs to be removed and recreated via gpinitstandby). This will
propagate the wal_level change. Before starting up the mirrors and
standby master, the hot_standby=on configuration needs to be set on
the standby master and mirrors (if you did it with existing mirrors,
could probably just restart the mirrors). Once everything is back up,
you should now be able to connect to the standby master and dispatch
to the HA mirrors.

TODO:
1. Make it easier to set up hot standby. Currently, it's very painful
to do all the configuration changes.
2. New DDL does not get applied on the standby master. Possibly
because the DTM is not initialized until after crash recovery is
completed and standby master is always in continuous recovery?
3. Make gp_segment_configuration flat file more viable. This will
allow remote replicas to be their own cluster which will allow DR
sites to be usable for read-only queries.