Allow quoted identifiers in gp_dist_random() input
The gp_dist_random pseudo function is special in that it takes in a
string literal which happens to be a relation name. The parsing logic
used on the string literal simply splits on the first dot and treats
the two tokens as schema and table. That works for common cases but
other cases such as those involving special characters can have issues
(e.g. a schema with a dot in its name would not split as expected). To
cover all cases, allow the schema and table to be quote identifiers so
that we can use regular Postgres parsing logic when constructing the
RangeVar.