WIP: GPDB replication slot API for finer control
The single function in this WIP commit allows a user to update the
replication slot restart_lsn manually instead of depending on
walsender/walreceiver interactions. This allows external programs to
use replication slots to keep WAL from being deleted until no longer
needed.

This is implemented as an extension because it's too dangerous to be a
default catalog function.