Fix psql calls in gpcopy test when using .psqlrc file
After the Postgres 9.6 merge into GPDB master, the .psqlrc file is
always parsed if the -X flag is not given to psql. This will make any
psql call to also print out the .psqlrc file parsing. Adding the -X
flag will make it always work as expected for this test.