Do not wait for synchronous replication if QD WAL sender is not for HA
In GPDB, the coordinator segment does not follow the usual rules for
synchronous replication. It just assumes that all WAL streaming
connections are synchronous no matter if the synchronous_standby_names
GUC is set or not (to allow the coordinator to skip blocking if the
standby coordinator crashes). To allow asynchronous connections to the
coordinator segment, we should check for the is_for_gp_walreceiver
flag that we label our High Availability mirror connections with. If
it's not set to true, we should not consider the WAL stream to be
synchronous.