Do not read .psqlrc when getting system locale for isolation2 stringsub
After the Postgres 9.6 merge, the .psqlrc file is always parsed if the
-X flag is not given to psql. In the isolation2 test suite, psql is
run to get the system locale but does it without the -X flag. If the
user has stuff in their .psqlrc file, the gp_collation isolation2 test
will fail due to a bad string substitution. To fix the issue, simply
just add the -X flag to the psql call.